1.4 update the extra credits. Instructions unclear
extra-credit-2 was quite obtuse on what it wanted!
He should have just explicitely stated the full intention on what he wanted us to do!
Anyway, in that lesson, when we capitalized our "custom component" name,
  we were supposed to do it so that we could replace our call to
  React.createElement with JSX -- and then also DO the replacement, now that it
  suddenly was possible!
He just said to rewrite the COMPONENT so it was easier to use... OK, I did...
  ...and the rest of the code was unchanged, and so the point
  (what progression???) was entirely dumb and unclear.
I hated this whole lesson until now. Like the what, & why of it.
What was the progression. What were we re-creating. It seemed disconnected.

I really think that instruction should be UPDATED in the COURSE !!!

Just tell us to rewrite our code, now that we can.
I thought we were just rewriting the Message function...
blah, blah, blah.