create empty extra-credit files
create empty extra-credit files for all remaining exercises.
Now once restart server, the platform can show and run any code I add
  to those files.
  (ie FIXES the platform, now can DO the extra credit
    without restarting the server EVERY time!
  The empty files should all have came with the course !!!,
    as the platform depends on those files already
    existing for each exercise set, when the exercise is started.
    and requires us to create the file and restart the server otherwise.
  )

Annoyingly, the course does not create files for the "extra credit".
And the way the course is designed, that means they do not show up in the
  "course" (platform).
  So, the server must be stopped and the file manually created every time.
  Then the SERVER MUST BE RESTARTED, each time an exercise/extra-credit
  file is created, in order for
  the course platform to be able to load it to show the file in the tab
  row, and to show the output of the exercise, and even to open the file
  up in an "isolated tab".  Unless the file already exists, the PLATFORM
  cannot embed the file into its interactive materials.
  And restarting the server takes Several MINUTES.

I get that he wanted the extra credit files to start out empty. That is fine.
But include the EMPTY FILES! As is, his UX is RE: extra credit, is ANNOYING!.
So, I created the remaining files so the platform has these empty files
available and able to be read into the platform.

NOTE TO SELF: FOR EACH MODULE DOWNLOADED:
  - change all tests to run on MY code, instead of his solution code!
  - create an extra credit file for EVERY extra credit exercise in the module.
    (I copied every extra credit solution file into a temp directory.
     Then opened and deleted the code from every file in that temp directory.
     saved and closed the file.
     Then moved all those files into my "exercise" directory, and deleted temp dir.
     This was the easiest way to capture every extra credit file name
     that will need to exist eventually.
     )